Conflicts:
	core/src/com/mygdx/sim/GameObjects/TrafficManager.java

+++ core/src/com/mygdx/sim/GameObjects/data/Edge.java
@@ -37,7 +37,7 @@ public class Edge implements Comparable<Edge> {
         this.speedLimit = speedLimit;
     }
 
-    public int getSpeedLimit(){
+    public double getSpeedLimit(){
         return speedLimit;
     }
